Side branch for building Silabs Unify Matter bridge
The Unify matter bridge source code is found in the Silicon Labs matter repo.
The Unify matter bridge relies on this branch as a library.

Repo for the Unify Matter bridge: https://github.com/SiliconLabs/matter

+++ applications/zpc/components/zwave/zwave_transports/s2/libs/zw-libs2/include/S2_external.h
@@ -125,7 +125,7 @@ void S2_dbg_printf(const char *str, ...);
 #ifdef ZIPGW
 unsigned long clock_time(void);
 #else
-uint16_t clock_time(void);
+uint32_t clock_time(void);
 #endif
 
 /**
