+++ applications/zpc/components/zwave/zwave_transports/s2/libs/zw-libs2/test/clock_time.c
@@ -3,7 +3,7 @@
 /**
 * Makes a fake System time available to LibS2.
 */
-uint16_t clock_time(void)
+uint32_t clock_time(void)
 {
   static uint16_t iTime;
